moved commands and replaced createGrade
